Refactor CDrop derivation to be explicitly derived and not derived when the CReprOf trait is derived.
Removes default implementation of do_drop method for CDrop trait.
